+++ packages/contracts/src/Configuration/ConfigurationValidator.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evolve\Contracts\Configuration;
+
+interface ConfigurationValidator
+{
+    public function validate(Configuration $configuration): void;
+}